Fix some leftovers

+++ src/PufferOracleV2.sol
@@ -58,7 +58,7 @@ contract PufferOracleV2 is IPufferOracleV2, AccessManaged {
         PUFFER_VAULT = vault;
         _totalNumberOfValidators = 927122; // Oracle will be updated with the correct value
         _epochNumber = 268828; // Oracle will be updated with the correct value
-        _setMintPrice(uint56(0.01 ether));
+        _setMintPrice(0.01 ether);
     }
 
     /**
